Upgrade Twine
The current version of Twine fails to publish packages due to errors in
the urllib package. Twine doesn't properly define its dependency on
urllib. Upgrading Twine fixes this issue.

This was tested by running CI on this branch and putting in an invalid
password to the Pypi repo. This resulted in auth failing but previously
the request was failing to be sent at all, so auth failure indicates we
have resolved the issue.